+++ task_list/application/domain/service/__init__.py
@@ -1,7 +1,8 @@
 from task_list.application.domain.service.add import AddService
 from task_list.application.domain.service.error import ErrorService
+from task_list.application.domain.service.execute import ExecuteService
 from task_list.application.domain.service.help import HelpService
 from task_list.application.domain.service.setdone import SetDoneService
 from task_list.application.domain.service.show import ShowService
 
-__all__ = ["AddService", "ErrorService", "HelpService", "SetDoneService", "ShowService"]
+__all__ = ["AddService", "ErrorService", "ExecuteService", "HelpService", "SetDoneService", "ShowService"]
